refactor(mentor): 컴포넌트별 폴더 생성 (디렉터리 구조 변경)
