+++ src/components/Header/Header.jsx
@@ -1,9 +1,9 @@
 import styles from "./Header.module.css";
-import logoPandaImg from "../assets/images/logo-panda.svg";
-import logoTxtImg from "../assets/images/logo-txt.svg";
-import avatarImg from "../assets/images/img-avatar.png";
-import Nav from "./Nav";
-import Avatar from "../components/Avatar";
+import logoPandaImg from "../../assets/images/logo-panda.svg";
+import logoTxtImg from "../../assets/images/logo-txt.svg";
+import avatarImg from "../../assets/images/img-avatar.png";
+import Nav from "../Nav";
+import Avatar from "../Avatar";
 
 const Header = () => {
   const handleAvatarClick = () => {
