+++ src/components/ItemList/ItemListRenderer.jsx
@@ -1,6 +1,6 @@
 import styles from "./ItemList.module.css";
-import ItemCard from "./ItemCard";
-import ItemCardSkeleton from "../ui/Skeletons/ItemCardSkeleton";
+import ItemCard from "../ItemCard";
+import ItemCardSkeleton from "../../ui/Skeletons/ItemCardSkeleton";
 
 const ItemListContent = ({ isLoading, isError, items, pageSize, listType }) => {
   const isLoadingError = !isLoading && isError;
