+++ src/ui/Input/InputSearch.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import styles from "./InputSearch.module.css";
-import searchImg from "../assets/images/ic-search.svg";
+import searchImg from "../../assets/images/ic-search.svg";
 
 const InputSearch = ({ className, placeholder }) => {
   const [value, setValue] = useState("");
